requirements.txt
python-dotenv==1.0.0
supabase==1.0.3
fastapi[all]==0.92.0